% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_job.R
\name{create_dir_structure}
\alias{create_dir_structure}
\title{Create a directory structure when saving the files}
\usage{
create_dir_structure(a, b, c)
}
\arguments{
\item{a}{The job title index}

\item{b}{The experience level index}

\item{c}{The location index}
}
\description{
The structure is:
  data/
  └── job title/
      └── experience level/
          └── location/
              └── file
}
