% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_attrs}
\alias{extract_attrs}
\title{A helper function to extract attributes}
\usage{
extract_attrs(i)
}
\arguments{
\item{i}{Input to the function}
}
\description{
A helper function to extract attributes
}
