% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_content}
\alias{extract_content}
\title{A helper function to extract content}
\usage{
extract_content(i, na.omit = T)
}
\arguments{
\item{i}{Input to the function}

\item{na.omit}{A boolean that specifies whether to remove NAs in the return}
}
\description{
A helper function to extract content
}
