% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_job_ad_metadata.R
\name{get_job_ad_metadata}
\alias{get_job_ad_metadata}
\title{Get the metadata for a page of LinkedIn job ad search results}
\usage{
get_job_ad_metadata(files, i = 1)
}
\arguments{
\item{files}{A vector of files with full system paths}

\item{i}{The index of the file in the list of files to extract information from. Defaults to 1. Because the scraped files have a lot of redundant information, the job ad metadata can be extracted from any one of the files. So this parameter usually doesn't need to be provided}
}
\value{
A data.frame containing the job_id, company, and URL link for all of the ads listed on the search page that was scraped.
}
\description{
It takes a page of search results and then for each job ad listed, it extracts the job ad description, and the criteria for the job ad.
}
