% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_job_description.R
\name{get_job_description}
\alias{get_job_description}
\title{Get the job description and job criteria from a scraped job ad file}
\usage{
get_job_description(files, i)
}
\arguments{
\item{files}{A vector of files with full system paths}

\item{i}{The index of the file in the list of files to extract information from.}
}
\value{
A list containing the description and job criteria
}
\description{
For each job ad, it will extract the job ad description, and the criteria for
the job ad.
}
