% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_job.R
\name{node_scrape}
\alias{node_scrape}
\title{Call NodeJS to scrape a LinkedIn search result}
\usage{
node_scrape(job_title, experience_level, location, sort = "R")
}
\arguments{
\item{job_title}{The title of the job to search for (e.g., 'data scientist')}

\item{experience_level}{The experience level to search for jobs (e.g., '1' corresponds to entry level jobs, '2' to Associate, '3' to Mid-senior, etc...)}

\item{location}{The name of the location to search in (e.g., 'New York City Metropolitan Area')}
}
\description{
When scraping, it saves html files for each job ad to the data/temp directory
}
