% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_chrome_path.R
\name{set_chrome_path}
\alias{set_chrome_path}
\title{Set the Chromium chrome.exe path for Node.js}
\usage{
set_chrome_path(path)
}
\arguments{
\item{path}{The path to Chromium chrome.exe}
}
\description{
Set the Chromium chrome.exe path for Node.js
}
