% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_page.R
\name{test_page}
\alias{test_page}
\title{View a scraped page in the browser}
\usage{
test_page(html)
}
\arguments{
\item{html}{The HTML to render}
}
\description{
This can be useful for debugging the output of a scrape job
}
